+++ R/covariate.R
@@ -206,6 +206,17 @@ sampleTimeVaryingCovariates <- function(object, armID, needsDV) {
 }
 
 #_______________________________________________________________________________
+#----                           loadFromJSON                                ----
+#_______________________________________________________________________________
+
+setMethod("loadFromJSON", signature=c("fixed_covariate", "json_element"), definition=function(object, json) {
+  object@name <- json@data$name
+  object@distribution <- loadFromJSON(new(json@data$distribution$type), 
+                                      JSONElement(json@data$distribution))
+  return(object)
+})
+
+#_______________________________________________________________________________
 #----                              sample                                   ----
 #_______________________________________________________________________________