+++ R/observations_set.R
@@ -18,9 +18,13 @@ setClass(
 #----                             getTimes                                  ----
 #_______________________________________________________________________________
 
+#' @param doseTimes times of the doses, only needed if a [DosingSchedule()] is referred to
 #' @rdname getTimes
-setMethod("getTimes", signature = c("observations_set"), definition = function(object) {
-  return(object@list %>% purrr::map(.f=~.x@times) %>% purrr::flatten_dbl() %>% unique() %>% base::sort())
+setMethod("getTimes", signature = c("observations_set"), definition = function(object, doseTimes=NULL) {
+  times <- object@list %>%
+    purrr::map(.f=~getTimes(.x, doseTimes=doseTimes)) %>%
+    purrr::flatten_dbl()
+  return(base::sort(unique(times)))
 })
 
 #_______________________________________________________________________________
@@ -28,7 +32,7 @@ setMethod("getTimes", signature = c("observations_set"), definition = function(o
 #_______________________________________________________________________________
 
 setMethod("show", signature=c("observations_set"), definition=function(object) {
-  times <- object %>% getTimes()
+  times <-  getTimes(object, doseTimes=NULL)
   cat(paste0("-> Obs. times: ", paste0(times, collapse=","), " (",
              times %>% length() , " observations in total)"))
 })