+++ R/repeated_schedule.R
@@ -30,6 +30,29 @@ setClass(
 )
 
 #_______________________________________________________________________________
+#----                         dosing_schedule class                         ----
+#_______________________________________________________________________________
+
+#' 
+#' Dosing schedule class.
+#' 
+#' @export
+setClass(
+  "dosing_schedule",
+  representation(
+  ),
+  contains="repeated_schedule"
+)
+
+#' Dosing schedule constructor.
+#' 
+#' @return a dosing schedule (schedule repeated at dose times)
+#' @export
+DosingSchedule <- function() {
+  return(new("dosing_schedule"))
+}
+
+#_______________________________________________________________________________
 #----                         cyclic_schedule class                         ----
 #_______________________________________________________________________________
 
@@ -114,6 +137,14 @@ setMethod("length", signature=c("cyclic_schedule"), definition=function(x) {
 #' @return a number
 setMethod("length", signature=c("repeat_at_schedule"), definition=function(x) {
   return(length(x@times))
+})
+
+#_______________________________________________________________________________
+#----                           loadFromJSON                                ----
+#_______________________________________________________________________________
+
+setMethod("loadFromJSON", signature=c("cyclic_schedule", "json_element"), definition=function(object, json) {
+  return(campsismod::mapJSONPropertiesToS4Slots(object, json))
 })
 
 #_______________________________________________________________________________